possible phase scenarios
It is Yugioh game rule that in one turn, player's phases start from Standby to Draw to Main1 Phase. The player can decide to go to Battle Phase OR end his turn. If the player chooeses to go Battle Phase, he can decide to go to 2nd Main Phase  OR end his turn.
If any player doesn't have a card to draw in DrawPhase, he loses the game.